+++ frontend/be-neutral/src/components/Dashboard.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Dashboard() {
+    return (
+        <div>
+            Dashboard here
+        </div>
+    )
+}
