+++ frontend/be-neutral/src/components/Other.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function Other() {
+    return (
+        <div>
+            Other page here
+        </div>
+    )
+}
